+++ CHANGES.rst
@@ -11,6 +11,12 @@
 
 - Added `~halotools.utils.sliding_conditional_percentile` function to `~halotools.utils` sub-package.
 
+- Added new `resample_x_to_match_y` function to `halotools.utils`.
+
+- Renamed old implementation of `conditional_abunmatch` to `conditional_abunmatch_bin_based`
+
+- Added new bin-free implementation of `conditional_abunmatch`.
+
 
 0.6 (2017-12-15)
 ----------------